Coalesce `Shape`s, move ad-hoc API to `AdHocShape` wrapper
I was initially confused why there are two (compatible) `Shape` struct definitions, and why their methods partially overlap, but with differences in the signature.

This PR doesn't resolve the duplication, but should make it clear what is deprecated and what is recommended, and what kind of API are consumers using.

It also makes Shape and AdHocShape fully interchangeable: AdHocShape can be created from Shape, destructures and (mutably) dereferences to it. That should make it possible to port usages incrementally.

Locally this diff is best viewed with `-C` / `--find-copies` so detect the ad-hoc API was just moved.
