+++ crates/opencascade/examples/box.rs
@@ -1,8 +1,8 @@
-use opencascade::Shape;
+use opencascade::adhoc::AdHocShape;
 
 pub fn main() {
-    let mut my_box = Shape::make_box(10.0, 10.0, 1.0);
-    let another_box = Shape::make_box(1.0, 1.0, 0.8);
+    let mut my_box = AdHocShape::make_box(10.0, 10.0, 1.0);
+    let another_box = AdHocShape::make_box(1.0, 1.0, 0.8);
 
     my_box.subtract(&another_box);
     my_box.chamfer_edges(0.07);
